pyproject.toml
[project]
name = "test-ingestion-infra"
version = "0.1.0"
description = "Testing poetry with a cdk repo"
authors = [
    {name = "cherry_dasika",email = "cherry1311@gmail.com"}
]
readme = "README.md"
requires-python = " ~=3.9"
dependencies = [
    "aws-cdk-lib (>=2.192.0,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "nox (>=2025.2.9,<2026.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
flake8 = "^7.2.0"
pre-commit = "^4.2.0"
nox = "^2025.2.9"

